Move indices alongside runtime
