+++ src/actors/builtin/storage_market/storage_market_actor_state.go
@@ -3,11 +3,11 @@ package storage_market
 import (
 	abi "github.com/filecoin-project/specs/actors/abi"
 	vmr "github.com/filecoin-project/specs/actors/runtime"
+	indices "github.com/filecoin-project/specs/actors/runtime/indices"
 	actor_util "github.com/filecoin-project/specs/actors/util"
 	filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
 	deal "github.com/filecoin-project/specs/systems/filecoin_markets/storage_market/storage_deal"
 	addr "github.com/filecoin-project/specs/systems/filecoin_vm/actor/address"
-	indices "github.com/filecoin-project/specs/systems/filecoin_vm/indices"
 )
 
 const epochUndefined = abi.ChainEpoch(-1)
