+++ src/actors/runtime/runtime.id
@@ -3,7 +3,7 @@ import abi "github.com/filecoin-project/specs/actors/abi"
 import addr "github.com/filecoin-project/specs/systems/filecoin_vm/actor/address"
 import exitcode "github.com/filecoin-project/specs/actors/runtime/exitcode"
 import filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
-import indices "github.com/filecoin-project/specs/systems/filecoin_vm/indices"
+import indices "github.com/filecoin-project/specs/actors/runtime/indices"
 import ipld "github.com/filecoin-project/specs/libraries/ipld"
 
 // Runtime is the VM's internal runtime object.
