+++ src/systems/filecoin_vm/runtime/impl/runtime.go
@@ -11,11 +11,11 @@ import (
 	initact "github.com/filecoin-project/specs/actors/builtin/init"
 	vmr "github.com/filecoin-project/specs/actors/runtime"
 	exitcode "github.com/filecoin-project/specs/actors/runtime/exitcode"
+	indices "github.com/filecoin-project/specs/actors/runtime/indices"
 	filcrypto "github.com/filecoin-project/specs/algorithms/crypto"
 	ipld "github.com/filecoin-project/specs/libraries/ipld"
 	actor "github.com/filecoin-project/specs/systems/filecoin_vm/actor"
 	addr "github.com/filecoin-project/specs/systems/filecoin_vm/actor/address"
-	indices "github.com/filecoin-project/specs/systems/filecoin_vm/indices"
 	msg "github.com/filecoin-project/specs/systems/filecoin_vm/message"
 	gascost "github.com/filecoin-project/specs/systems/filecoin_vm/runtime/gascost"
 	st "github.com/filecoin-project/specs/systems/filecoin_vm/state_tree"
